Split out dramRepl stuff in its on file
